+++ .nvmrc
@@ -1,2 +1,1 @@
 v6.10.3
-
