+++ package.json
@@ -6,8 +6,16 @@
   "repository": "git@github.com:JohnTheSixth/packbot.git",
   "author": "<john@burge.io>",
   "license": "MIT",
+  "dependencies": {
+    "axios": "^0.17.1"
+  },
   "devDependencies": {
-    "serverless": "^1.25.0"
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
+    "babel-loader": "^7.1.2",
+    "serverless": "^1.26.0",
+    "serverless-plugin-webpack": "^1.4.0",
+    "webpack": "^3.10.0"
   },
   "scripts": {
     "sls": "./node_modules/.bin/serverless"
